Migrate code into fitter module
